+++ tests/RedisConnectionTest.php
@@ -6,6 +6,7 @@ use yii\helpers\ArrayHelper;
 use yii\log\Logger;
 use yii\redis\Connection;
 use yii\redis\SocketException;
+use yiiunit\extensions\redis\support\ConnectionWithErrorEmulator;
 
 /**
  * @group redis
@@ -136,6 +137,32 @@ class ConnectionTest extends TestCase
 
         $this->assertTrue($db->ping());
         $this->assertCount(11, $logger->messages, 'log +1 ping command, and reconnection.'
+            . print_r(array_map(function($s) { return (string) $s; }, ArrayHelper::getColumn($logger->messages, 0)), true));
+    }
+
+    public function testConnectionTimeoutRetryWithFirstFail()
+    {
+        $logger = new Logger();
+        Yii::setLogger($logger);
+
+        $databases = TestCase::getParam('databases');
+        $redis = isset($databases['redis']) ? $databases['redis'] : [];
+        $db = new ConnectionWithErrorEmulator($redis);
+        $db->retries = 3;
+
+        $db->configSet('timeout', 1);
+        $this->assertCount(3, $logger->messages, 'log of connection and init commands.');
+
+        $this->assertTrue($db->ping());
+        $this->assertCount(4, $logger->messages, 'log +1 ping command.');
+
+        sleep(2);
+
+        // Set flag for emulate socket error
+        $db->isTemporaryBroken = true;
+
+        $this->assertTrue($db->ping());
+        $this->assertCount(10, $logger->messages, 'log +1 ping command, and two reconnections.'
             . print_r(array_map(function($s) { return (string) $s; }, ArrayHelper::getColumn($logger->messages, 0)), true));
     }