+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Yii Framework 2 redis extension Change Log
 2.0.17 under development
 ------------------------
 
-- no changes in this release.
+- Enh #176: Fix reconnect logic bug, add `protected function sendRawCommand()` (ilyaplot)
 
 
 2.0.16 October 04, 2021